Move methods from RootController to UsersController
