+++ server/routes/root_router.js
@@ -25,11 +25,6 @@ router.route('/')
 
 /**
  * Routes
- *   POST /users/auth
- */
-
-/**
- * Routes
  *   PUT /users/:id/mediator-upgrade
  */
